Removed AbsInteractor (inlined into DoInteractor)
