+++ domain/src/main/java/com/teamagam/gimelgimel/domain/base/interactors/DoInteractor.java
@@ -5,9 +5,10 @@ import com.teamagam.gimelgimel.domain.base.subscribers.DummyObserver;
 import io.reactivex.Observable;
 import io.reactivex.observers.ResourceObserver;
 
-public abstract class DoInteractor<T> extends AbsInteractor<T> {
+public abstract class DoInteractor<T> implements Interactor {
 
   private final ThreadExecutor threadExecutor;
+  private ResourceObserver<T> mObserver;
 
   protected DoInteractor(ThreadExecutor threadExecutor) {
     this.threadExecutor = threadExecutor;
@@ -21,5 +22,15 @@ public abstract class DoInteractor<T> extends AbsInteractor<T> {
 
   protected ResourceObserver<T> getObserver() {
     return new DummyObserver<>();
+  }
+
+  public final void execute() {
+    mObserver = buildObservable().subscribeWith(getObserver());
+  }
+
+  public final void unsubscribe() {
+    if (!mObserver.isDisposed()) {
+      mObserver.dispose();
+    }
   }
 }
